Build improvements and changes to retrieving build config from Python
Replace `python_config` with `python_lib_dir`. This option can be used when
`python` is not executable on the local host. The path, which should either
be the Python `lib` directory or the Python build directory, will be
searched for the `_sysconfigdata_*.py` file to extract build vars.

Add `tools/python_config.py` with `get_python_config_vars()` for retrieving
the build vars from `python` or `_sysconfigdata_*.py`.

Use a set instead of a list for sources to avoid duplicates

Add `CLASS_VIRTUAL_CALL_TABLE_SIZE` define for setting the virtual call
table to a smaller value when needed by the build system.