Fixup reference counts when returning and receiving returned values
Semantics are a bit tricky here as Python will only hold a single Godot
reference. When writing to a pointer to return a refcounted object the
ref count must be incremented, as the local reference may go out of scope
before the reader of the pointer has a chance to increment. The combination
of these two facts requires that careful fixup of ref counts is done Python
side when dealing with return via pointer.

Passing and receiving arguments is not affected by this, as the creator
of the argument will hold a reference the entire duration of the call, so
the callee can safely increment the ref count themselves (only once per
object in the case of Python).

This commit adds code to determine when the return type is itself an object
reference type (and not merely an object pointer that could receive a
reference counted object) and adjusts reference counts accordingly.